Split hashing & rendering
