+++ lib/pmdtester.rb
@@ -22,6 +22,7 @@ require_relative 'pmdtester/builders/diff_report/configerrors'
 require_relative 'pmdtester/builders/diff_report/errors'
 require_relative 'pmdtester/builders/diff_report_builder'
 require_relative 'pmdtester/builders/project_builder'
+require_relative 'pmdtester/builders/project_hasher'
 require_relative 'pmdtester/builders/pmd_report_builder'
 require_relative 'pmdtester/builders/liquid_html_builder'
 require_relative 'pmdtester/builders/rule_set_builder'
