+++ resources/project_diff_report.html
@@ -1,6 +1,6 @@
 <html lang="en">
 <head>
-    <title>{{title}}</title>
+    <title>Diff report for {{project_name}}</title>
 
     <link rel="stylesheet" type="text/css" href="css/bootstrap.min.css"/>
     <link rel="stylesheet" type="text/css" href="css/datatables.min.css"/>
